refactor(telegram): move incoming-media helpers to media.rs, verbatim (#471 phase 1)
Third mechanical cut: caption merging, tmp file download/saving,
recent-tmp-file lookup and the size-capped getFile wrapper move to a
dedicated media.rs. Pure move, pub(crate) visibility bumps, glob
re-export keeps all paths stable.
